requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlmodel==0.0.19
sqlalchemy==2.0.30
PyMuPDF==1.24.5
EbookLib==0.18
watchdog==4.0.1
Pillow==10.3.0
python-multipart==0.0.9
aiofiles==23.2.1
lxml==5.2.2
html5lib==1.1
